Throw exception on out of memory in dynamic buffer
js_dbuf_init users that check the return value of functions like
dbuf_put should be able to assume that an OOM exception is pending
when the function returns an error. No callers currently do so but
I am going to start doing that in the not too distant future.

js_dbuf_init takes a JSContext pointer, strongly suggesting it calls
js_realloc, but in fact it called js_realloc_rt, which does *not*
raise OOM exceptions.

Call js_realloc and move js_dbuf_realloc closer to js_dbuf_init for
legibility reasons.